+++ eras/conway/impl/test/Test/Cardano/Ledger/Conway/BinarySpec.hs
@@ -2,11 +2,10 @@
 
 module Test.Cardano.Ledger.Conway.BinarySpec (spec) where
 
-import Cardano.Ledger.Alonzo.TxAuxData
 import Cardano.Ledger.Conway
 import Test.Cardano.Ledger.Common
 import Test.Cardano.Ledger.Conway.Arbitrary ()
 import Test.Cardano.Ledger.Core.Binary as UpgradeSpec
 
 spec :: Spec
-spec = specUpgrade @Conway @AlonzoTxAuxData True
+spec = specUpgrade @Conway True